Rename coroutines package to adapters
