+++ sandwich/src/main/kotlin/com/skydoves/sandwich/adapters/ApiResponseCallDelegate.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.skydoves.sandwich.coroutines
+package com.skydoves.sandwich.adapters
 
 import com.skydoves.sandwich.ApiResponse
 import retrofit2.Call
