+++ sandwich/src/main/kotlin/com/skydoves/sandwich/adapters/CallDelegate.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.skydoves.sandwich.coroutines
+package com.skydoves.sandwich.adapters
 
 import com.skydoves.sandwich.SandwichInitializer
 import okhttp3.Request
