+++ sandwich/src/main/kotlin/com/skydoves/sandwich/adapters/DataSourceCallDelegate.kt
@@ -14,25 +14,26 @@
  * limitations under the License.
  */
 
-package com.skydoves.sandwich.coroutines
+package com.skydoves.sandwich.adapters
 
 import com.skydoves.sandwich.DataSource
+import com.skydoves.sandwich.ResponseDataSource
 import retrofit2.Call
-import retrofit2.CallAdapter
-import java.lang.reflect.Type
+import retrofit2.Callback
+import retrofit2.Response
 
 /**
  * @author skydoves (Jaewoong Eum)
  *
- * CoroutinesDataSourceCallAdapter is an coroutines call adapter for creating [DataSource] from service method.
- *
- * request API network call asynchronously and returns [DataSource].
+ * DataSourceCallDelegate is a delegate [Call] proxy for handling and transforming normal generic type [T]
+ * as [DataSource] that wrapping [T] data from the network responses.
  */
-internal class DataSourceCallAdapter constructor(
-  private val responseType: Type
-) : CallAdapter<Type, Call<DataSource<Type>>> {
+internal class DataSourceCallDelegate<T>(proxy: Call<T>) : CallDelegate<T, DataSource<T>>(proxy) {
 
-  override fun responseType(): Type = responseType
+  override fun enqueueImpl(callback: Callback<DataSource<T>>) {
+    val responseDataSource = ResponseDataSource<T>().combine(proxy, null)
+    callback.onResponse(this@DataSourceCallDelegate, Response.success(responseDataSource))
+  }
 
-  override fun adapt(call: Call<Type>): Call<DataSource<Type>> = DataSourceCallDelegate(call)
+  override fun cloneImpl() = DataSourceCallDelegate(proxy.clone())
 }
